pyproject.toml
[project]
name = "ml-project-group"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "albumentations>=1.4.0",
    "matplotlib>=3.7.0",
    "numpy>=1.23.5",
    "opencv-python-headless>=4.10.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "sahi>=0.11.15",
    "seaborn>=0.13.0",
    "supervision>=0.23.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "tqdm>=4.66.0",
    "ultralytics>=8.3.0",
    # Florence-2 Teacher dependencies
    "transformers>=4.40.0",
    "timm>=1.0.0",
    "einops>=0.8.0",
    "kaggle>=1.8.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]

# ADD THIS BLOCK
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true  # This is the magic keyword. It stops uv from searching here by accident.

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchvision = [{ index = "pytorch" }]
